THAT MOCKOON IN THERE THAT FUZZY IN THERE IT WORKS IT ADDING ONLY TICKETS THAT A USER HAS.
but yeah just moved the whole fuzzy implementation into a seperate python script.

Added parameters and imported required packages to make a call to an outside API (Mockoon in this case)

Modified the Parsing and Fuzzy similarity score.

overrided the similarity score if the venue was in parathesis as this would be a logical string manipulation outside the scope of the Levenshtein distance algorithm.

And added debugging to help me, which I am just going to keep in there.